+++ components/pm-subscription-handler/pmsh_service/mod/pmsh_utils.py
@@ -15,7 +15,9 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 # ============LICENSE_END=====================================================
+import os
 import uuid
+import schema
 from os import getenv
 from threading import Timer
 
@@ -25,7 +27,6 @@ from onaplogging.mdcContext import MDC
 from requests.auth import HTTPBasicAuth
 from tenacity import wait_fixed, stop_after_attempt, retry, retry_if_exception_type
 
-import mod.network_function
 from mod import logger
 from mod.subscription import Subscription
 
@@ -41,6 +42,7 @@ def mdc_handler(function):
         kwargs['request_id'] = request_id
         kwargs['invocation_id'] = invocation_id
         return function(*args, **kwargs)
+
     return decorator
 
 
@@ -58,6 +60,15 @@ class MySingleton(object):
         return type(clz.__name__, (MySingleton,), dict(clz.__dict__))
 
 
+def _load_sub_schema_from_file():
+    try:
+        with open(os.path.join(os.path.dirname(__file__), 'sub_schema')) as sub:
+            schema_string = sub.read()
+        return eval(schema_string)
+    except OSError as err:
+        logger.error(f'Failed to read sub schema file: {err}', exc_info=True)
+
+
 class AppConfig:
     INSTANCE = None
 
@@ -73,8 +84,9 @@ class AppConfig:
         self.streams_publishes = conf['config'].get('streams_publishes')
         self.operational_policy_name = conf['config'].get('operational_policy_name')
         self.control_loop_name = conf['config'].get('control_loop_name')
+        self.sub_schema = schema.Schema(_load_sub_schema_from_file())
         self.subscription = Subscription(**conf['policy']['subscription'])
-        self.nf_filter = mod.network_function.NetworkFunctionFilter(**self.subscription.nfFilter)
+        self.nf_filter = None
 
     def __new__(cls, *args, **kwargs):
         if AppConfig.INSTANCE is None:
@@ -102,6 +114,23 @@ class AppConfig:
         except Exception as e:
             logger.error(f'Failed to get config from CBS: {e}', exc_info=True)
             raise ValueError(e)
+
+    def validate_sub_schema(self):
+        """
+        Validates schema of PMSH subscription
+
+        Raises:
+            SchemaError: If the PMSH subscription schema is invalid
+        """
+        sub_data = self.subscription.__dict__
+        self.sub_schema.validate(sub_data)
+        nf_filter = sub_data["nfFilter"]
+        for filter_name in nf_filter:
+            if len(nf_filter[filter_name]) > 0:
+                break
+        else:
+            raise schema.SchemaError("At least one filter within nfFilter must not be empty")
+        logger.debug("Subscription schema is valid.")
 
     def refresh_config(self):
         """